Split DeclarationSlices module in two.
